bpftune: unexpected exit on suspend
bpftune was exiting on suspend; the reason is we get an -EINTR while
polling.  However because it is not a SIGINT/SIGTERM, we should not
exit the polling routine for this error; the signal handling logic
will do this for the signals we care about.

Tested via

	$ systemctl suspend

bpftune service/program keeps running where it previously exited.

Reported by: Tungsten842 (https://github.com/Tungsten842)
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>