tcp_conn_tuner: move to using reinforcement learing approach
Previously, congestion control algoritm selection was done via a simple
method of looking for packet loss on a per-connection basis and applying
BBR if a limit was surpassed.

This approach has been replaced by using a reinforcement learning
based algorithm which seeks to choose the congestion control algorithm
that approaches optimal performance for the link; minimum RTT and
delivery rate tuned to fill the pipe.  See bpftune-tcp-conn(8) for
details.

The behaviour under loss is similar to before - BBR is selected -
but it is selected based upon its optimal behaviour rather than
via a hand-coded algorithm.  In normal network conditions, different
algorithms can be selected; choices available currently are

- cubic
- BBR
- htcp
- dctcp

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>