tcp_buffer_tuner: fix load failure for 6.9+
TCP buffer tuner was failing to load on 6.9 and later.

CO-RE relocation for sk_userlocks in struct sock was the cause;
the problem is sk_userlocks changes from a bitfield to a plain
u8 and the vmlinux.h that was generated had the old representation.
We can use a Linux version check to guard check of sk_userlocks.

Reported-by: HippieMitch <https://github.com/HippieMitch>
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>