fix global namespace disable
when we change a sysctl external to bpftune, the aim is to disable the
auto-tuning for the associated namespace. however in the global ns
we were disabling tuning for global and non-global namespaces via
bpftuner_fini().  Ensure we disable global namespace only, and that we
only send events for tuners that are enabled and - if namespaced -
if the namespace is enabled also.

Reported by: Roger Knobbe (https://github.com/rknobbe)
Signed-off-by: Alan Maguire <alan.maguire@oracle.com>