libbpftune: quiet clang analyzer complaints about unfreed memory
Because the clang analyzer does not look into libraries, it
thinks the BPF skeleton objects calloc()ed in BPF skeletons are
not freed because a wrapper function is used to free them.  Work
around this to convince the analyzer the memory is freed since we
cannot change the bpftool-generated skeletons.

Signed-off-by: Alan Maguire <alan.maguire@oracle.com>